ANDROID: fuse-bpf: Use fuse_bpf_args in uapi
fuse_args is not suitable for use in the uapi - it is not stable, and
contains internal pointers. Replace with stable equivalent.

The end_offset values are currently unused and unset, but will be used
in a follow up patch by the verifier.

Test: fuse_test, atest ScopedStorageDeviceTest pass
Bug: 202785178
Signed-off-by: Daniel Rosenberg <drosen@google.com>
[ bengris32: Resolve merge conflict on android12-5.10 ]
Change-Id: Ic1c12f9706aeae233cc30a0d68ed2533030e485b
(cherry picked from commit d8620caae64e314c5a34e4eb36e99f9fe813011f)